requirements.txt
click==6.7
Django==1.8.17
django-appconf==1.0.2
django-bootstrap3==8.2.1
django-bootstrap3-sass==3.3.2.2
django-compressor==2.1.1
django-crispy-forms==1.6.1
django-libsass==0.7
Flask==0.12
itsdangerous==0.24
Jinja2==2.9.5
libsass==0.12.3
MarkupSafe==1.0
mysqlclient==1.3.10
olefile==0.44
Pillow==4.0.0
ply==3.10
rcssmin==1.0.6
requests==2.13.0
rjsmin==1.0.12
six==1.10.0
slimit==0.8.1
Werkzeug==0.12.1
